fix: address PR review comments (security, correctness, performance)
backend:
- auth.py: select signing key by kid from JWKS (was passing entire JWKS)
- auth.py: enable audience verification with COGNITO_CLIENT_ID
- auth.py: use hmac.compare_digest for constant-time secret comparison
- auth.py: pre-fetch JWKS/origin-secret at startup via FastAPI lifespan
- main.py: fix allow_credentials=False (True + allow_origins=* crashes)
- main.py: handle empty CLOUDFRONT_DOMAIN (strip + fallback to *)
- models.py: income/expense/transfer int → float (multi-currency support)
- routers: move boto3 resource to module level (connection reuse)

frontend:
- main.ts: remove unused getIdToken import
- main.ts: deterministic txId via SHA-256 hash (prevent duplicate uploads)
- index.html: fix label for="from-filter" (was month-filter)

infra:
- lambda: remove invalid layer_arn arg from data source; use var.lwa_layer_arn
- lambda: add COGNITO_CLIENT_ID env var + cognito_client_id input variable
- cloudfront: remove 403 SPA fallback (would swallow API 403 responses)
- prod/lambda: pass cognito_client_id from cognito dependency

+++ backend/app/models.py
@@ -12,7 +12,7 @@ class Transaction(BaseModel):
     category: str
     subcategory: str
     shop: str
-    income: int
-    expense: int
-    transfer: int
+    income: float
+    expense: float
+    transfer: float
     aggregation: str
